Move clientID fetch logic into own class
