+++ CorePayments/src/main/java/com/paypal/android/corepayments/analytics/AnalyticsService.kt
@@ -12,6 +12,7 @@ import java.util.*
 
 class AnalyticsService internal constructor(
     private val deviceInspector: DeviceInspector,
+    private val clientIDAPI: ClientIDAPI,
     private val configuration: CoreConfig,
     private val http: Http,
     private val httpRequestFactory: HttpRequestFactory,
@@ -21,6 +22,7 @@ class AnalyticsService internal constructor(
     constructor(configuration: CoreConfig, context: Context, orderID: String) :
         this(
             deviceInspector = DeviceInspector(context),
+            clientIDAPI = ClientIDAPI(configuration),
             configuration = configuration,
             http = Http(),
             httpRequestFactory = HttpRequestFactory(),
@@ -33,7 +35,7 @@ class AnalyticsService internal constructor(
     fun sendAnalyticsEvent(name: String) {
         GlobalScope.launch {
             try {
-                val clientID = fetchCachedOrRemoteClientID()
+                val clientID = clientIDAPI.fetchCachedOrRemoteClientID()
                 performEventRequest(name, clientID)
             } catch (e: PayPalSDKError) {
                 Log.d(
@@ -47,12 +49,13 @@ class AnalyticsService internal constructor(
     internal suspend fun performEventRequest(name: String, clientID: String) {
         val timestamp = System.currentTimeMillis()
 
+        // TODO: rename sessionID to orderID
         val analyticsEventData = AnalyticsEventData(
             clientID,
             configuration.environment.name.lowercase(),
             name,
             timestamp,
-            sessionId,
+            orderID,
             deviceInspector.inspect()
         )
         val httpRequest = httpRequestFactory.createHttpRequestForAnalytics(analyticsEventData)
@@ -61,47 +64,5 @@ class AnalyticsService internal constructor(
         if (!response.isSuccessful) {
             throw APIClientError.clientIDNotFoundError(response.status, "")
         }
-    }
-
-    /**
-     * Retrieves the merchant's clientID either from the local cache, or via an HTTP request if not cached.
-     * @return Merchant clientID.
-     */
-    @Throws(PayPalSDKError::class)
-    suspend fun fetchCachedOrRemoteClientID(): String {
-        API.clientIDCache.get(configuration.accessToken)?.let { cachedClientID ->
-            return cachedClientID
-        }
-
-        val apiRequest = APIRequest("v1/oauth2/token", HttpMethod.GET)
-        val httpRequest =
-            httpRequestFactory.createHttpRequestFromAPIRequest(apiRequest, configuration)
-        val response = http.send(httpRequest)
-        val correlationID = response.headers["Paypal-Debug-Id"]
-        if (response.isSuccessful) {
-            val clientID = parseClientId(response.body, correlationID)
-            API.clientIDCache.put(configuration.accessToken, clientID)
-            return clientID
-        }
-
-        throw APIClientError.serverResponseError(correlationID)
-    }
-
-    @Throws(PayPalSDKError::class)
-    private fun parseClientId(responseBody: String?, correlationID: String?): String {
-        if (responseBody.isNullOrBlank()) {
-            throw APIClientError.noResponseData(correlationID)
-        }
-        val json = PaymentsJSON(responseBody)
-        val clientID = json.optString("client_id")
-        if (clientID.isNullOrBlank()) {
-            throw APIClientError.dataParsingError(correlationID)
-        }
-        return clientID
-    }
-
-    companion object {
-        private val sessionId = UUID.randomUUID().toString()
-        // val clientIDCache = LruCache<String, String>(10)
     }
 }
