+++ PayPalNativePayments/src/main/java/com/paypal/android/paypalnativepayments/PayPalNativeCheckoutClient.kt
@@ -1,11 +1,7 @@
 package com.paypal.android.paypalnativepayments
 
 import android.app.Application
-import com.paypal.android.corepayments.API
-import com.paypal.android.corepayments.CoreConfig
-import com.paypal.android.corepayments.CoreCoroutineExceptionHandler
-import com.paypal.android.corepayments.PayPalSDKError
-import com.paypal.android.corepayments.APIClientError
+import com.paypal.android.corepayments.*
 import com.paypal.android.corepayments.analytics.AnalyticsService
 import com.paypal.checkout.PayPalCheckout
 import com.paypal.checkout.approve.OnApprove
@@ -26,8 +22,10 @@ import kotlinx.coroutines.launch
 /**
  * Use this client to checkout with PayPal.
  */
+// TODO: - remove api
 class PayPalNativeCheckoutClient internal constructor (
     private val application: Application,
+    private val clientIDAPI: ClientIDAPI,
     private val coreConfig: CoreConfig,
     private val returnUrl: String,
     private val api: API,
@@ -46,6 +44,7 @@ class PayPalNativeCheckoutClient internal constructor (
      */
     constructor(application: Application, coreConfig: CoreConfig, returnUrl: String) : this(
         application,
+        ClientIDAPI(coreConfig),
         coreConfig,
         returnUrl,
         API(coreConfig, application)
@@ -87,7 +86,7 @@ class PayPalNativeCheckoutClient internal constructor (
         orderID = request.orderID
         CoroutineScope(dispatcher).launch(exceptionHandler) {
             try {
-                val clientID = api.fetchCachedOrRemoteClientID()
+                val clientID = clientIDAPI.fetchCachedOrRemoteClientID()
                 val config = CheckoutConfig(
                     application = application,
                     clientId = clientID,
@@ -163,7 +162,7 @@ class PayPalNativeCheckoutClient internal constructor (
     }
 
     private fun notifyOnCancel() {
-        api.sendAnalyticsEvent("paypal-native-payments:canceled")
+        analyticsService?.sendAnalyticsEvent("paypal-native-payments:canceled")
         listener?.onPayPalCheckoutCanceled()
     }
 }
